refactor: 디렉터리 정리 (#22)
* refactor: api route와 헷갈리지 않도록 app 외부 api 디렉터리 이름을 apis로 변경

* refactor: context 절대경로 추가 및 커스텀 훅 이름 변경

* refactor: components 디렉터리 layout, document, common으로 분리

* fix: global layout에서 manifest 분리

* style: apple-touch-icon 퍼블릭 디렉터리로 이동

* refactor: page params import로 변경

* fix: page revalidate 시간 때문에 발생하던 빌드 오류 해결
