+++ client/src/app/api/put-document/route.ts
@@ -1,10 +1,10 @@
 'use server';
 
-import {PostDocumentContent} from '@api/document';
+import {PostDocumentContent} from '@apis/document';
 
 import {WikiDocument} from '@type/Document.type';
 import {ENDPOINT} from '@constants/endpoint';
-import {requestPut} from '@utils/http';
+import {requestPut} from '@apis/http';
 import {NextRequest, NextResponse} from 'next/server';
 import {revalidateTag} from 'next/cache';
 import {CACHE} from '@constants/cache';
