+++ client/src/app/wiki/[title]/edit/page.tsx
@@ -1,16 +1,16 @@
 'use client';
 
-import PostHeader from '@components/Write/PostHeader';
-import {DocumentWriteContextProvider, useDocumentWriteContextProvider} from '../../../../context/DocumentWriteContext';
-import TitleInputField from '@components/Write/TitleInputField';
-import TuiEditor from '@components/MarkdownEditor';
-import RelativeSearchTerms from '@components/SearchTerms/RelativeSearchTerms';
+import PostHeader from '@components/document/Write/PostHeader';
+import {DocumentWriteContextProvider, useDocumentWriteContext} from '@context/DocumentWriteContext';
+import TitleInputField from '@components/document/Write/TitleInputField';
+import TuiEditor from '@components/document/MarkdownEditor';
+import RelativeSearchTerms from '@components/common/SearchTerms/RelativeSearchTerms';
 import {useParams} from 'next/navigation';
 import {useRelativeSearchTerms} from '@app/wiki/post/useRelativeSearchTerms';
 import {useGetDocumentByTitle} from '@hooks/fetch/useGetDocumentByTitle';
 
 const EditPage = () => {
-  const {editorRef, initialContents} = useDocumentWriteContextProvider();
+  const {editorRef, initialContents} = useDocumentWriteContext();
   const {top, left, titles, onClick, showRelativeSearchTerms} = useRelativeSearchTerms({editorRef});
 
   return (
