+++ client/src/components/document/TOC/AttachTocClickHandler.tsx
