+++ client/src/components/document/layout/DocumentFooter.tsx
