+++ client/src/components/document/layout/MobileDocumentHeader.tsx
@@ -1,4 +1,4 @@
-import Button from '@components/Button';
+import Button from '@components/common/Button';
 import {URLS} from '@constants/urls';
 import Link from 'next/link';
