--- client/src/constants/dummy.ts
@@ -1,64 +1,0 @@
-const TEXT = `
-# 회의록
-## 해야할 것
-### backend
-- 뭐라는건지 모르겠음 아니 전 프엔이라;; 자료 주세요
-- 이미지는 프론트에서 주는거 들어서 알아서 하겠음
-- s3 등 아직 생각안하는 중
-- 환경설정만 완료
-- 기획팀아 기획서 내놔
-
-### 3/18 ~ 3/24
-- 기획: 잘 되고있는지 24일 내로 각 팀의 개발이 완료될것 같은지
-- 프론트: 퍼블리싱 -> 기타 기능들 붙여 나감
-- 백엔드: api 설계
-
-### 이미지 처리 방법
-- 프론트에서 toast ui로 이미지 받는 방식 시도해 보고 처리 방식 고려하기
-- cloudinary
-	- 백엔드 부담 최소화
-	- 리사이징 해줄 수 있음
-	- 비용은 아직 모름
-- notion을 db처럼 사용해서 image upload
-- 별도 서버는 network 문제 있음
-
-### 문서 제목!!!!
-닉네임 (기수)
-- 토다리(6기)
-
-### 방향성
-- SETP1에서 로그인 기능 없음 편집자, 작성자 어떻게 할건지
-- 1. 임시적으로 작성자 input 란 만들기
-	- input 입력한 닉네임이 6기 크루에 있는지
-
-- 유저들에게 어떻게 유용함을 줄 수 있을 것인지?
-- 캠퍼스 마감하는 법
-- 외에도 사용자들의 팁 등
-- 알쓸크잡 내용 있으면 좋을듯
-> 모든 크루에게 프로필을 제공하자!
-
-### 프로필 필수요소
-- 닉네임
-- 기수
-- 과정
-- 생일 (월/일)
-- MBTI
-
-## milestone
-### step1 ~ 4/5
-작성 및 편집 + 프로필
-로그 - 전체 문서 때려박기
-페이지 url, 참조 + 정확한 검색
-사진용량 제한 - 모듈 간단함
-최근 변경 문서
-### step2
-회원가입 - oauth
-2차 인증 - 기수, 분야, 닉네임, 출입증사진 혹은 그때가서 슬랙을하든 지랄을하든 알아서
-로그인 - oauth
-어드민
-신고
-참조한 문서
-검색
-`;
-
-export default TEXT;
