+++ client/src/hooks/fetch/useGetDocumentByTitle.ts
@@ -1,6 +1,6 @@
 'use client';
 
-import {getDocumentByTitle} from '@api/document';
+import {getDocumentByTitle} from '@apis/document';
 import {useFetch} from '@hooks/useFetch';
 import {WikiDocument} from '@type/Document.type';
 import {useCallback} from 'react';
