+++ client/src/hooks/mutation/usePutDocument.ts
@@ -1,10 +1,10 @@
 'use client';
 
-import {PostDocumentContent} from '@api/document';
+import {PostDocumentContent} from '@apis/document';
 import {URLS} from '@constants/urls';
 import useMutation from '@hooks/useMutation';
 import {WikiDocument} from '@type/Document.type';
-import {requestPut} from '@utils/http';
+import {requestPut} from '@apis/http';
 import {useRouter} from 'next/navigation';
 
 export const usePutDocument = () => {
